refactor: bottom bar 페이지 라우팅 관련 수정
