+++ lib/component/bottom_navigation_bar.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class MainBottomNavigationBar extends StatelessWidget {
+  final int selectedIndex;
+  final ValueChanged<int> onItemTapped;
+
+  const MainBottomNavigationBar({
+    Key? key,
+    required this.selectedIndex,
+    required this.onItemTapped,
+    }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ClipRRect(
+      borderRadius: BorderRadius.only(
+          topLeft: Radius.circular(20), topRight: Radius.circular(20)),
+      child: BottomNavigationBar(
+        type: BottomNavigationBarType.fixed,
+        currentIndex: selectedIndex,
+        onTap: onItemTapped,
+        items: [
+          BottomNavigationBarItem(icon: Icon(Icons.home), label: "Home"),
+          BottomNavigationBarItem(icon: Icon(Icons.report), label: "Report"),
+          BottomNavigationBarItem(icon: Icon(Icons.group), label: "Group"),
+          BottomNavigationBarItem(icon: Icon(Icons.favorite), label: "Match"),
+          BottomNavigationBarItem(icon: Icon(Icons.people), label: "MyPage"),
+        ],
+      ),
+    );
+  }
+}
