+++ lib/screen/page_report.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class PageReport extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('Report'),
+    );
+  }
+}
